requirements.txt
asgiref==3.3.4
autopep8==1.5.7
backcall==0.2.0
branca==0.4.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cloudinary==1.26.0
colorama==0.4.4
cryptography==3.4.7
cycler==0.10.0
decorator==5.0.9
Deprecated==1.2.12
dj-database-url==0.5.0
Django==3.2.4
django-ckeditor==6.1.0
django-cloudinary-storage==0.3.0
django-heroku==0.3.1
django-js-asset==1.2.2
djongo==1.3.6
dnspython==2.1.0
folium==0.12.1
gcloud==0.18.3
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
jwcrypto==0.9.1
jws==0.1.3
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
newsapi-python==0.2.6
numpy==1.21.0
oauth2client==4.1.3
pandas==1.2.5
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
plotly==4.14.3
prompt-toolkit==3.0.18
protobuf==3.17.0
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.1
Pygments==2.9.0
pymongo==3.11.4
pyparsing==2.4.7
Pyrebase4==4.4.3
python-dateutil==2.8.1
python-dotenv==0.18.0
python-jwt==2.0.1
pytz==2021.1
pywin32==301
pyzmq==22.1.0
requests==2.25.1
requests-toolbelt==0.9.1
retrying==1.3.3
rsa==4.7.2
six==1.16.0
sqlparse==0.2.4
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
whitenoise==5.2.0
wrapt==1.12.1
